+++ packages/camera/camera_platform_interface/test/method_channel/method_channel_camera_test.dart
@@ -586,6 +586,7 @@ void main() {
               'cameraId': cameraId,
               'maxVideoDuration': null,
               'enableStream': false,
+              'videoOutputPath': null,
             },
           ),
         ]);