+++ packages/camera/camera_web/lib/src/camera_web.dart
@@ -439,7 +439,11 @@ class CameraPlugin extends CameraPlatform {
   }
 
   @override
-  Future<void> startVideoRecording(int cameraId, {Duration? maxVideoDuration}) {
+  Future<void> startVideoRecording(
+    int cameraId, {
+    Duration? maxVideoDuration,
+    String? videoOutputPath,
+  }) {
     // Ignore maxVideoDuration, as it is deprecated.
     return startVideoCapturing(VideoCaptureOptions(cameraId));
   }