+++ packages/camera/camera_windows/lib/camera_windows.dart
@@ -187,9 +187,13 @@ class CameraWindows extends CameraPlatform {
   }
 
   @override
-  Future<void> startVideoRecording(int cameraId, {Duration? maxVideoDuration}) async {
+  Future<void> startVideoRecording(
+    int cameraId, {
+    Duration? maxVideoDuration,
+    String? videoOutputPath,
+  }) async {
     // Ignore maxVideoDuration, as it is unimplemented and deprecated.
-    return startVideoCapturing(VideoCaptureOptions(cameraId));
+    return startVideoCapturing(VideoCaptureOptions(cameraId, videoOutputPath: videoOutputPath));
   }
 
   @override
@@ -199,7 +203,7 @@ class CameraWindows extends CameraPlatform {
     }
 
     // Currently none of `options` is supported on Windows, so it's not passed.
-    await _hostApi.startVideoRecording(options.cameraId);
+    await _hostApi.startVideoRecording(options.cameraId, options.videoOutputPath);
   }
 
   @override