+++ packages/camera/camera_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.10.12
+
+* Adds support for custom video output path in video recording.
+
 ## 0.10.11
 
 * Adds `setJpegImageQuality` for controlling JPEG compression quality.